requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
pillow==11.2.1
python-dotenv==0.21.1
qrcode==8.2
requests==2.28.2
requests-oauthlib==2.0.0
six==1.16.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==1.26.20
websocket-client==1.4.2
Werkzeug==3.1.3
xumm-sdk-py==1.0.4
